dialyzer: Remove undocumented v1 solver
This was kept around to test the v2 solver under the latter's
development, but has remained untested ever since and that which is
not tested does not work.

Since dialyzer's test suites cover less than we'd like and w rely
mostly on testing in-the-wild, we have little choice but to remove
oxbow code like this it if we're going to have any hope of
maintaining it.