erl_types: Ignore opaque type arguments
TL;DR: Opaque types are rather convoluted, this commit removes one
of the hairiest parts.

----

It is not commonly known that opaque types are invariant on their
type parameters: an opaque type `foo:bar(gurka | gaffel)` _IS NOT_
a subtype of `foo:bar(atom())`, nor is it a supertype of
`foo:bar(gurka)`. For an opaque type to be compatible with
another instance of the same type, its type parameters have to be
exactly equal.

This may seem a bit strict but is required for correctness:
while it's trivial to see that one can pass an `[integer()]` to a
function expecting `[number()]` since the list type is covariant on
the element type, we don't have a way to express whether an opaque
type parameter is covariant, contravariant, or invariant. Nor can
we infer it beyond toy cases.

The easiest way to keep the type system sound in the face of that
is to treat all type parameters as invariant. It's perhaps a bit
surprising but there's nothing wrong with it. So far so good.

Then we hit this marvel in `gb_sets`:

```erlang
-spec add(Element, Set1) -> Set2 when
      Set1 :: set(Element),
      Set2 :: set(Element).
```

A strict reading of this says that it accepts a set containing a
singleton `Element`, and produces a set containing `Element`
alone. If we were to pass a set produced by this function to one
that expects `gb_sets:set(atom())`, we would get a type error
unless `Element` had already collapsed to `atom()`. (Other funny
consequences are left as an exercise to the reader)

Yet that works at the moment. Huh?!

1. dialyzer just throws away type variables in specs, replacing
   them with `term()` and removing all relations between them.

   Trying to fix that is how I noticed this wonderful quirk.
2. dialyzer considers `term()` in type variables to be quasi-equal
   to all other types. That is, `gb_sets:set(term())` is okay in
   functions that expect (say) `gb_sets:set(atom())` and vice
   versa.

The former is bad but enough but the latter is where things go off
the rails. We may pass `any()`d and more strictly typed opaques
around interchangeably, but since parameter types are still
invariant there's no way for us to tell which is the subtype and
which is the supertype! This effectively breaks the type lattice
and sends dialyzer into infinite loops often enough that adding a
test case to this commit is rather pointless: you can hardly even
begin building a PLT before hitting it.

Since we can't do anything about this without inventing a new
notation and requiring people to rewrite their specs with it, I've
decided to ignore type parameters altogether in this commit. It
makes dialyzer less useful, but I believe it's worth it to preserve
the notion that dialyzer is never wrong.