beam_call_types: Operate on type() instead of normal_type()
This is needed to keep the type information consistent in an
upcoming commit that optimizes record updates. It will introduce
a psuedo-instruction that does several updates at once, which gets
converted either to a specialized instruction or a series of
`setelement/3` calls during code generation.

As the `setelement/3` is taken through `beam_call_types`, and that
only dealt with normalized types prior to this commit, its type
information would be less precise and could cause validation to
fail.