erts: Fix stopping 'long_timers_test' of other versions
If the long timers test lingers in the background during tests,
and can even persist after tests have been aborted. To deal with
this, the test tries to stop previous nodes, but does so with a
local fun which fails if the OTP version has changed in the
meantime, for example if rebasing something to an earlier version.

To fix this, we'll use an external fun instead so that we'll be
capable of stopping nodes of any version from here on.