inet_config: Tolerate atoms as `inetrc` kernel argument
Many RabbitMQ users provided atoms instead of strings as the
`inetrc` file path because of an error in their documentation. This
was tolerated by accident until atom filename support was removed
from erl_prim_loader.

We did the latter for good reason, but tolerating this specifically
in inet_config is no big deal.